backend/pyproject.toml
[project]
name = "kafi-chatbot-backend"
version = "0.1.0"
description = "Standalone backend for Kafi Chatbot using Qwen3-8B"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pyyaml",
    "python-dotenv",
    "transformers>=4.48.3",
    "torch==2.9.1",
    "bitsandbytes>=0.49.1",
    "accelerate>=1.12.0",
    "pandas",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
